Fix XSS issue in the online editor
It was possible to embed HTML into parser output in the online editor
using e.g the following grammar together with input "foo":

  start = "foo" { return "<b>baz</b>"; }

The problem was caused by incomplete HTML escaping by the jsDump
library. I overlooked a note in the post at http://bit.ly/bWpBNC when
writing the original code:

  Note that the script won't escape all the text, only the parts added
  by the script (indentation, new lines, dumped nodes), it won't alter
  original strings, etc.

I solved the problem by switching off HTML escaping in jsDump and
rewriting the code to use <pre> for the output so the plain-text
formatting jsDump does is displayed correctly.